Search heavier for objects that are referenced
In some cases git push from a client returned with a missing blob or
a missing tree.  This is caused by a pack that contains a
reference to a tree or blob that in fact is reachable through one
of the advertised references but not directly in the tree of the
advertised reference but in one of its parents / grandparents trees.

This fix starts a new ObjectWalk at the advertised references and
tries to incrementally find objects that are not found otherwise.
When the object is found the ObjectWalk is suspended and resumed
if needed for another object. Before an object is searched via
the ObjectWalk a db.hasObject() triggers fast fail when an object
is not in the repository at all, However this alone is not sufficient
as this could break Gerrit access control policies.